server: Re-organize app file structure and create Groups app entrypoint
Create a new Node.js application file that serves Groups-related
endpoints only, with minimal setup.

This necessarily splits the previous application file into multiple
parts that can be re-used by the new file. Put these in a new directory
for routing.
