+++ package.json
@@ -15,6 +15,7 @@
     "build": "./build.sh",
     "lint": "eslint --ext .js,.jsx .",
     "server": "node server.js",
+    "groups": "node server.js --app ./src/groupsApp.js",
     "start": "npm run server",
     "//test": "echo 'NODE_OPTIONS is required until this issue is resolved: https://github.com/facebook/jest/issues/9430 '",
     "test": "NODE_OPTIONS='--experimental-vm-modules' jest --roots ./test/",
