+++ src/app.js
@@ -1,32 +1,12 @@
-import sslRedirect from 'heroku-ssl-redirect';
-
-import nakedRedirect from 'express-naked-redirect';
-import express from 'express';
-import compression from 'compression';
-import cors from 'cors';
-import stream from 'stream';
-import { promisify } from 'util';
-
-/* XXX TODO: Replace promisify() with require("stream/promises") once we
- * upgrade to Node 15+.
- *   -trs, 5 Nov 2021
- */
-const streamFinished = promisify(stream.finished);
-
 const CANARY_ORIGIN = process.env.CANARY_ORIGIN;
 
-import { PRODUCTION } from './config.js';
-import * as utils from './utils/index.js';
-import { addAsync } from './async.js';
-import { Forbidden, NotFound, Unauthorized } from './httpErrors.js';
-import * as authn from './authn/index.js';
+import { NotFound } from './httpErrors.js';
 import * as endpoints from './endpoints/index.js';
-import { AuthzDenied } from './exceptions.js';
-import { replacer as jsonReplacer } from './json.js';
-import * as middleware from './middleware.js';
 import * as redirects from './redirects.js';
+import * as routing from "./routing/index.js";
+import { setupApp } from "./routing/setup.js";
 import * as sources from './sources/index.js';
-import { uri } from './templateLiterals.js';
+import * as utils from './utils/index.js';
 
 const {
   setSource,
@@ -44,80 +24,30 @@ const {
 } = endpoints.sources;
 
 const {
-  optionsGroup,
-  getGroupLogo,
-  putGroupLogo,
-  deleteGroupLogo,
-  getGroupOverview,
-  putGroupOverview,
-  deleteGroupOverview,
-} = endpoints.groups;
-
-const {
   CoreSource,
   CoreStagingSource,
   CommunitySource,
   UrlDefinedSource,
-  GroupSource,
 } = sources;
 
-const jsonMediaType = type => type.match(/^application\/(.+\+)?json$/);
+const {
+  auspice,
+  charon,
+  errors,
+  groups,
+  openid,
+  staticSite,
+} = routing;
 
 
-/* Express boilerplate.
- */
-const app = addAsync(express());
-
-app.set("json replacer", jsonReplacer);
-
-app.locals.production = PRODUCTION;
-app.locals.gatsbyDevUrl = PRODUCTION ? null : process.env.GATSBY_DEV_URL;
-
-// In production, trust Heroku as a reverse proxy and Express will use request
-// metadata from the proxy.
-if (PRODUCTION) app.enable("trust proxy");
-
-app.use(sslRedirect()); // redirect HTTP to HTTPS
-app.use(compression()); // send files (e.g. res.json()) using compression (if possible)
-app.use(nakedRedirect({reverse: true})); // redirect www.nextstrain.org to nextstrain.org
-app.use(middleware.rejectParentTraversals);
-
-
-/* Setup a request-scoped context object for passing arbitrary request-local
- * data between route and param middleware and route handlers.  Akin to
- * app.locals or res.locals.
- */
-app.use((req, res, next) => {
-  req.context = {};
-
-  // Set the app's origin centrally so other handlers can use it
-  //
-  // We can trust the HTTP Host header (req.hostname) because we're always
-  // running behind name-based virtual hosting on Heroku.  A forged Host header
-  // will be rejected by Heroku and never make it to us.
-  req.context.origin = PRODUCTION
-    ? `${req.protocol}://${req.hostname}`
-    : `${req.protocol}://${req.hostname}:${req.app.get("port")}`;
-
-  next();
-});
-
-
-/* Authentication (authn) setup.
+/* Express app setup.
  *
- * Routes:
- *   GET /login
- *   GET /logged-in
- *   GET /logout
+ * Includes authn routes:
+ *   /login
+ *   /logged-in
+ *   /logout
  */
-authn.setup(app);
-
-
-/* CORS.
- *
- * After authn so it doesn't apply to those routes.
- */
-app.use(middleware.allowPublicReadOnlyCors);
+const app = setupApp();
 
 
 /* Canary.
@@ -146,38 +76,15 @@ redirects.setup(app);
 
 
 /* Charon API used by Auspice.
+ *
+ * Routes:
+ *   /charon/getAvailable
+ *   /charon/getDataset
+ *   /charon/getNarrative
+ *   /charon/getSourceInfo
+ *   /charon/*
  */
-if (!PRODUCTION) {
-  // allow cross-origin from the gatsby dev server
-  app.use("/charon", cors({origin: 'http://localhost:8000'}));
-}
-
-app.routeAsync("/charon/getAvailable")
-  .getAsync(endpoints.charon.getAvailable);
-
-app.routeAsync("/charon/getDataset")
-  .getAsync(
-    endpoints.charon.setSourceFromPrefix,
-    endpoints.charon.setDatasetFromPrefix,
-    endpoints.charon.canonicalizeDatasetPrefix,
-    endpoints.charon.getDataset);
-
-app.routeAsync("/charon/getNarrative")
-  .getAsync(
-    endpoints.charon.setSourceFromPrefix,
-    endpoints.charon.setNarrativeFromPrefix,
-    endpoints.charon.getNarrative);
-
-app.routeAsync("/charon/getSourceInfo")
-  .getAsync(
-    endpoints.charon.setSourceFromPrefix,
-    endpoints.charon.getSourceInfo);
-
-app.routeAsync("/charon/*")
-  .all((req) => {
-    utils.warn(`(${req.method}) ${req.url} has not been handled / has no handler`);
-    throw new NotFound();
-  });
+charon.setup(app);
 
 
 /* Core datasets and narratives
@@ -304,105 +211,25 @@ app.routeAsync("/fetch/:authority/*")
 
 
 /* Groups
+ *
+ * Routes:
+ *   /groups/:groupName
+ *   /groups/:groupName/settings
+ *   /groups/:groupName/settings/logo
+ *   /groups/:groupName/settings/overview
+ *   /groups/:groupName/settings/members
+ *   /groups/:groupName/settings/roles
+ *   /groups/:groupName/settings/roles/:roleName/members
+ *   /groups/:groupName/settings/roles/:roleName/members/:username
+ *   /groups/:groupName/settings/*
+ *   /groups/:groupName/narratives
+ *   /groups/:groupName/narratives/*
+ *   /groups/:groupName/*
+ *   /whoami
+ *   /users
+ *   /users/:name
  */
-
-app.use("/groups/:groupName",
-  setSource(req => new GroupSource(req.params.groupName)),
-
-  // Canonicalize the Group name.
-  (req, res, next) => {
-    const restOfUrl = req.url !== "/" ? req.url : "";
-
-    const canonicalName = req.context.source.group.name;
-    const canonicalUrl = uri`/groups/${canonicalName}` + restOfUrl;
-
-    return req.params.groupName !== canonicalName
-      ? res.redirect(canonicalUrl)
-      : next();
-  });
-
-app.routeAsync("/groups/:groupName")
-  /* sendGatsbyPage("groups/:groupName/index.html") should work, but it
-   * renders wrong for some reason that's not clear.
-   */
-  .getAsync(endpoints.static.sendGatsbyEntrypoint)
-;
-
-app.use("/groups/:groupName/settings",
-  endpoints.groups.setGroup(req => req.params.groupName));
-
-app.routeAsync("/groups/:groupName/settings")
-  .getAsync(endpoints.static.sendGatsbyEntrypoint);
-
-app.routeAsync("/groups/:groupName/settings/logo")
-  .getAsync(getGroupLogo)
-  .putAsync(putGroupLogo)
-  .deleteAsync(deleteGroupLogo)
-  .optionsAsync(optionsGroup)
-;
-
-app.routeAsync("/groups/:groupName/settings/overview")
-  .getAsync(getGroupOverview)
-  .putAsync(putGroupOverview)
-  .deleteAsync(deleteGroupOverview)
-  .optionsAsync(optionsGroup)
-;
-
-app.routeAsync("/groups/:groupName/settings/members")
-  .getAsync(endpoints.groups.listMembers);
-
-app.routeAsync("/groups/:groupName/settings/roles")
-  .getAsync(endpoints.groups.listRoles);
-
-app.routeAsync("/groups/:groupName/settings/roles/:roleName/members")
-  .getAsync(endpoints.groups.listRoleMembers);
-
-app.routeAsync("/groups/:groupName/settings/roles/:roleName/members/:username")
-  .getAsync(endpoints.groups.getRoleMember)
-  .putAsync(endpoints.groups.putRoleMember)
-  .deleteAsync(endpoints.groups.deleteRoleMember)
-;
-
-app.route("/groups/:groupName/settings/*")
-  .all(() => { throw new NotFound(); });
-
-// Avoid matching "narratives" as a dataset name.
-app.routeAsync("/groups/:groupName/narratives")
-  .getAsync((req, res) => res.redirect(uri`/groups/${req.params.groupName}`));
-
-app.routeAsync("/groups/:groupName/narratives/*")
-  .all(setNarrative(req => req.params[0]))
-  .getAsync(getNarrative)
-  .putAsync(putNarrative)
-  .deleteAsync(deleteNarrative)
-  .optionsAsync(optionsNarrative)
-;
-
-app.routeAsync("/groups/:groupName/*")
-  .all(setDataset(req => req.params[0]))
-  .getAsync(getDataset)
-  .putAsync(putDataset)
-  .deleteAsync(deleteDataset)
-  .optionsAsync(optionsDataset)
-;
-
-
-/* Users
- */
-app.routeAsync("/whoami")
-  .getAsync(endpoints.users.getWhoami);
-
-/* For requests for text/html, the first /whoami implementation returned bare
- * bones HTML dynamic on the logged in user.  This was later replaced by a
- * redirect to /users/:name (for a logged in session) or /users (if no one
- * logged in), which then rendered a Gatsby page.  However, the /users/:name
- * form was purely aesthetic: :name was not used for routing at all.  Instead
- * of keeping up this fiction, have /whoami render the Gatsby page directly,
- * at least until we _actually_ implement user profile pages.
- *   -trs, 4 Oct 2021
- */
-app.route(["/users", "/users/:name"])
-  .get((req, res) => res.redirect("/whoami"));
+groups.setup(app);
 
 
 /* CLI convenience endpoints, e.g. downloads of release assets.
@@ -454,158 +281,31 @@ app.route("/schemas/*")
   .all((req, res, next) => next(new NotFound()));
 
 
-/* OpenID Connect 1.0 configuration.  Retrieved by Nextstrain CLI to
- * discovery necessary authentication details.
+/* OpenID Connect 1.0 configuration.
  *
- * <https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata>
+ * Routes:
+ *   /.well-known/openid-configuration
  */
-app.routeAsync("/.well-known/openid-configuration")
-  .getAsync(endpoints.openid.providerConfiguration);
+openid.setup(app);
 
 
-/* Auspice HTML pages and assets.
+/* Auspice.
  *
- * Auspice hardcodes URL paths that start with /dist/… in its Webpack config,
- * so we must use that prefix here too.
+ * Routes:
+ *   /dist/*
+ *   /edit/narratives
  */
-app.route("/dist/*")
-  .all(endpoints.static.auspiceAssets, (req, res, next) => next(new NotFound()));
+auspice.setup(app);
 
-/* Auspice has a special /edit/narratives route -
- * It is not backed by a dataset, and only exists for GET requests
+
+/* static-site (Gatsby).
  */
-app.routeAsync("/edit/narratives")
-  .getAsync(endpoints.static.sendAuspiceEntrypoint);
+await staticSite.setup(app);
 
-/* Gatsby static HTML pages and other assets.
- *
- * Any URLs matching Gatsby's static files will be handled here, e.g.
- * static-site/public/influenza/index.html will be served for /influenza.
- *
- * When a Gatsby dev server is in use, assets are proxied from the dev server
- * instead of served straight from disk.
+
+/* Error handling.
  */
-if (app.locals.gatsbyDevUrl) {
-  const {createProxyMiddleware} = await import("http-proxy-middleware");
-
-  // WebSocket endpoint
-  app.get("/socket.io/", createProxyMiddleware(app.locals.gatsbyDevUrl, { ws: true }));
-
-  /* This is the end of the line in gatsbyDevUrl mode, as the proxy middleware
-   * doesn't fallthrough on 404 and even if it did the Gatsby dev server
-   * returns a 200 Ok for missing pages.
-   */
-  app.use(createProxyMiddleware(app.locals.gatsbyDevUrl));
-} else {
-  app.use(endpoints.static.gatsbyAssets);
-}
-
-
-/* Everything else gets 404ed.
- */
-app.use((req, res, next) => next(new NotFound()));
-
-
-/* Error handler
- */
-app.useAsync(async (err, req, res, next) => {
-  /* XXX TODO: Replace calls to Express' next() with calls to our own custom
-   * finalhandler (the library Express uses) function configured with an
-   * "onerror" handler that does what we want with regard to logging.
-   *   -trs, 1 Oct 2021
-   */
-  if (res.headersSent) {
-    utils.verbose("Headers already sent; using Express' default error handler");
-    return next(err);
-  }
-
-  /* Read the entire request body (discarding it) if the request might have a
-   * body and wasn't made with Expect: 100-continue, or if it was and we wrote
-   * a 100 Continue response but then ended up here.  This ensures that the
-   * request is finished being sent before we return a (error) response, which
-   * some clients require (such as Heroku's routing/proxy layer and the Python
-   * "requests" package, but notably not curl).
-   */
-  const mayHaveBody = !["GET", "HEAD", "DELETE", "OPTIONS"].includes(req.method);
-
-  if (mayHaveBody && (!req.expectsContinue || res.wroteContinue)) {
-    const reqFinished = streamFinished(req);
-    req.unpipe();
-    req.resume();
-    await reqFinished;
-  }
-
-  res.vary("Accept");
-
-  /* "Is this request browser-like?"  Checking for explicit inclusion of
-   * "text/html" is an imperfect heuristic, but still useful enough and doesn't
-   * require user-agent matching, which seems more fraught and more opaque.
-   *
-   * Note that we don't check req.accepts("text/html"), because that'll match
-   * wildcard Accept values which are sent by ~every client.
-   *   -trs, 25 Jan 2022
-   */
-  const isBrowserLike = req.accepts().includes("text/html");
-
-  /* Handle our authorization denied errors differently depending on if the
-   * request is authenticated or not and if the client is browser-like or not.
-   *
-   * The intended audience for the redirect is humans following bookmarks,
-   * browser history, or other saved links, which will only ever be GET (and
-   * _maybe_ HEAD).
-   *
-   * An additional redirect condition on navigation (vs. background request)
-   * would also be nice, but I can't find any good heuristic for that.
-   * The following seems ideal:
-   *
-   *    const isNavigation = req.headers['sec-fetch-mode'] === "navigate";
-   *
-   * but it is not supported by Safari (macOS or iOS).
-   *   -trs, 25 Jan 2022
-   */
-  if (err instanceof AuthzDenied) {
-    if (!req.user) {
-      if (["GET", "HEAD"].includes(req.method) && isBrowserLike) {
-        utils.verbose(`Redirecting anonymous user to login page from ${req.originalUrl}`);
-        req.session.afterLoginReturnTo = req.originalUrl;
-        return res.redirect("/login");
-      }
-      err = new Unauthorized(err.message);
-    } else {
-      err = new Forbidden(err.message);
-    }
-  }
-
-  /* Browser-like clients get JSON if they explicitly ask for it (regardless of
-   * priority, and including our custom +json types) and all non-browser like
-   * clients get JSON.
-   */
-  if (req.accepts().some(jsonMediaType) || !isBrowserLike) {
-    utils.verbose(`Sending ${err} error as JSON`);
-    return res.status(err.status || err.statusCode || 500)
-      .json({
-        error: err.message || String(err),
-        ...(
-          !PRODUCTION
-            ? {stack: err.stack}
-            : {}
-        ),
-      })
-      .end();
-  }
-
-  if (err instanceof NotFound) {
-    /* A note about routing: if the current URL path (i.e. req.path) matches a
-     * a page known to Gatsby (e.g. via the Gatsby page's "path" or "matchPath"
-     * properties), then Gatsby will perform client-side routing to load that
-     * page even though we're serving a static page (404.html) here.
-     */
-    return await endpoints.static.sendGatsby404(req, res);
-  }
-
-  utils.verbose(`Sending ${err} error as HTML with Express' default error handler`);
-  return next(err);
-});
+errors.setup(app);
 
 
 export default app;
