+++ src/routing/index.js
@@ -1,0 +1,15 @@
+import * as auspice from './auspice.js';
+import * as charon from './charon.js';
+import * as errors from "./errors.js";
+import * as groups from './groups.js';
+import * as openid from './openid.js';
+import * as staticSite from './staticSite.js';
+
+export {
+  auspice,
+  charon,
+  errors,
+  groups,
+  openid,
+  staticSite,
+};
